test: refactor sequencer inbox
